+++ src/Infrastructure/Ci/System/Github/Credentials/GithubActionsCredentialsMapper.php
@@ -3,7 +3,7 @@
 namespace ArtARTs36\MergeRequestLinter\Infrastructure\Ci\System\Github\Credentials;
 
 use ArtARTs36\MergeRequestLinter\Domain\CI\Authenticator;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Ci\Credentials\TokenAuthenticator;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Ci\Credentials\HeaderAuthenticator;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\CI\AuthenticatorMapper;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\Configuration\ConfigValueTransformer;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Http\Exceptions\InvalidCredentialsException;
@@ -30,6 +30,6 @@ class GithubActionsCredentialsMapper implements AuthenticatorMapper
             ));
         }
 
-        return TokenAuthenticator::bearer($this->value->tryTransform($credentials['token']));
+        return HeaderAuthenticator::bearer($this->value->tryTransform($credentials['token']));
     }
 }